Replace templates with driver function
