+++ lib/activator.js
@@ -16,7 +16,6 @@ getObjectProperty = function(object, property) {
 DEFAULTS = {
 	model: {find: function(user,cb){cb("uninitialized");}, save: function(id,data,cb){cb("uninitialized");}, generate: null },
 	transport: "smtp://localhost:465/activator.net/",
-	templates: __dirname+'/templates',
 	resetExpire: 60,
 	proto: "https://",
 	emailProperty: "email",
@@ -245,7 +244,7 @@ module.exports = {
 	init: function (config) {
 		model = config.user || DEFAULTS.model;
 		transport = config.transport || DEFAULTS.transport;
-		templates = config.templates || DEFAULTS.templates;
+		templates = config.templates || function(type,lang,callback){callback(null);};
 		resetExpire = config.resetExpire || DEFAULTS.resetExpire;
 		proto = config.protocol || DEFAULTS.proto;
 		mailer = smtp(transport,templates, config.styliner || DEFAULTS.styliner);
@@ -319,5 +318,8 @@ module.exports = {
 			_.extend(req.activator,{code:code,message:message});
 			next();
 		});
-	}	
+	},
+	templates: {
+		file: require('./filesdriver')
+	}
 };
