+++ lib/mailer.js
@@ -9,7 +9,7 @@ module.exports = function(url, templates, styliner){
 			if (subject && (text||html)) {
 				sendmail(from,to,subject,text,html,attachments,callback);
 			} else {
-				callback("missingmailfile");
+				callback("missingmailtemplate "+type);
 			}
 		});		
 	};
